pyproject.toml
[project]
name = "grafana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
